fix: use default exports for humps
The humps library provides uses default exports to export it's functionality.

This becomes a problem for ESM bundling as tsup just includes the following line in the bundle:

import { camelizeKeys, decamelize, decamelizeKeys } from 'humps';

Running in ESM mode causes failures because named exports are not present in the humps module.